# OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: Project Management Orchestrator - Revised
  version: 2.0.0
  description: |
    Unified API with proper organization isolation and role-based access control.
    
    **Authentication**: All endpoints require Bearer token except /register and /login
    **Authorization**: Enforced at endpoint level based on user role
    **Data Isolation**: Users can only access data from their organization
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://api.example.com
    description: Production server

# ============================================
# AUTHENTICATION & ORGANIZATION MANAGEMENT
# ============================================
paths:
...
  /api/v1/labs:
    post:
      tags: [Labs]
      summary: Create lab in organization
      description: |
        ⚠️ Only admins can create labs.
        Creates lab in user's organization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description, head_id]
              properties:
                name:
                  type: string
                description:
                  type: string
                domain:
                  type: string
                head_id:
                  type: string
                  format: uuid
                  description: Must be org member
      responses:
        '201':
          description: Lab created
        '403':
          description: Only admins can create labs
        '400':
          description: Head must be in organization

    get:
      tags: [Labs]
      summary: List labs in user's organization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of labs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /api/v1/labs/{lab_id}:
    get:
      tags: [Labs]
      summary: Get lab details
      description: Returns lab details if user is in the lab's organization
      security:
        - bearerAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lab details
        '403':
          description: Not authorized to view this lab
  /api/v1/labs/{lab_id}/researchers:
    post:
      tags: [Labs]
      summary: Add researcher to lab
      description: |
        ⚠️ Only lab head or admin can add researchers.
        Links existing org user to lab as researcher.
      security:
        - bearerAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, field]
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: Must be org member
                field:
                  type: string
                  description: Research field/specialization
                status:
                  type: string
                  enum: [pending, active]
                  default: pending
      responses:
        '201':
          description: Researcher added to lab
        '403':
          description: Only lab head or admin can add researchers
        '400':
          description: User not in organization or already in lab

    get:
      tags: [Labs]
      summary: List researchers in lab
      security:
        - bearerAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of researchers
...
  /api/v1/projects:
    post:
      tags: [Projects]
      summary: Create project
      description: |
        ⚠️ Only admin or manager can create projects.
        Creates project in user's organization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                lab_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional lab link
      responses:
        '201':
          description: Project created
        '403':
          description: Only admin or manager can create projects
        '400':
          description: Lab not in organization (if lab_id provided)

    get:
      tags: [Projects]
      summary: List projects in user's organization
      security:
        - bearerAuth: []
      parameters:
        - name: lab_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by lab
      responses:
        '200':
          description: List of projects
...
  /api/v1/dashboard:
    get:
      tags: [Dashboard]
      summary: Get unified dashboard
      description: Returns personalized dashboard based on user role
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    description: User info and role
                  organization:
                    type: object
                    description: Organization details
                  projects:
                    type: array
                    description: User's projects
                  tasks:
                    type: array
                    description: Tasks assigned to user
                  activities:
                    type: object
                    description: Recent activities
                  performance:
                    type: object
                    description: Performance metrics
                  team_stats:
                    type: object
                    description: Team statistics (if manager/admin)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...